Fix hashing problem. Things are much faster now.

+++ Icosahedron/IcoCollection.cs
@@ -13,7 +13,7 @@ namespace Icosahedron
         public IcoCollection(int level)
         {
             this.level = level;
-            points = IcoCommon.VerticiesInLevel(level);
+            points = (int)IcoCommon.VerticiesInLevel(level);
             values = new T[points];
         }
